requirements.txt
python-dotenv
streamlit
langchain>=0.1.0
langchain-community
langchain-google-genai
google-generativeai
faiss-cpu
pypdf
openpyxl
